+++ NEWS.md
@@ -2,6 +2,7 @@
 
 ## Changes in version 1.1.7
 - Add `mz,FragmentViews-methods` [2018-02-01].
+- Remove internal `fragmentMass` and `fragmentNames` functions [2018-02-22].
 
 ## Changes in version 1.1.6
 - Rotate fragment labels (vertical orientation) in `plot` [2018-01-17].